Failed to properly calculate the cache position when the size of the cache exceeded 2GB. Now the limit is "size_t" which is 64-bit.

+++ mdl2/pthread/mdl.c
@@ -1106,7 +1106,7 @@ void mdlFree(MDL mdl,void *p) {
 */
 static void *getArrayElement(void *vData,int i,int iDataSize) {
     char *pData = vData;
-    return pData + i*iDataSize;
+    return pData + (size_t)i * (size_t)iDataSize;
     }
 
 void mdlSetCacheSize(MDL mdl,int cacheSize) {
